feat: show open zaak validation errors on UI
Solves [PZ-7314]
